Move DBFieldEncryption classes to the `entity/encryption` folder
