+++ rpc/wallet.ts
@@ -20,7 +20,7 @@ import { MetricsReporter } from '../src/MetricsReporter';
 import { LoadedContracts } from './contracts';
 import { RPCMethod, RPCNamespace } from './decorators';
 import { AwsPrivateKeyDBFieldEncryption } from "../src/shipchain/AwsPrivateKeyDBFieldEncryption";
-import { PrivateKeyDBFieldEncryption } from "../src/encryption/PrivateKeyDBFieldEncryption";
+import { PrivateKeyDBFieldEncryption } from "../src/entity/encryption/PrivateKeyDBFieldEncryption";
 
 const loadedContracts = LoadedContracts.Instance;
 const metrics = MetricsReporter.Instance;
