+++ src/__tests__/contracts.ts
@@ -20,7 +20,7 @@ import 'mocha';
 import { createConnection } from 'typeorm';
 import { Wallet } from '../entity/Wallet';
 import { Contract, Version, Project, Network } from '../entity/Contract';
-import { PrivateKeyDBFieldEncryption } from "../encryption/PrivateKeyDBFieldEncryption";
+import { PrivateKeyDBFieldEncryption } from "../entity/encryption/PrivateKeyDBFieldEncryption";
 
 const utils = require('../local-test-net-utils');
 const GETH_NODE = process.env.GETH_NODE || 'http://localhost:8545';
